fix broken assimp usage on windows
VCPKG changed the layout of the CMake file from assimp.
Earlier ASSIMP_LIBRARIES_RELEASE pointed to the release library.
Now the variable does not exists anymore and ASSIMP_LIBRARIES leverage
the usage of CMake build in optimized/debug options. With this patch
IPPP is able to build with latest version of vcpkg.

Signed-off-by: Tobias Kohlbau <tobias@kohlbau.de>